<IB_DESIGNABLE and IBInspectable>

+++ FoodTracker/RatingControlView.h
@@ -11,5 +11,5 @@
 IB_DESIGNABLE
 @interface RatingControlView : UIStackView
 @property (nonatomic, assign) IBInspectable CGSize startSize;
-@property (nonatomic, assign) IBInspectable NSInteger startCount;
+@property (nonatomic, assign) IBInspectable int startCount;
 @end
